Create app.py
This version of the code organizes files of different types like  images, documents, videos, executables, installers, presentations and torrents. 

the files are sorted based on the sizes insider their respective folder like small, medium and large sizes. This is the first version of the code will update it as I get new ideas!